refactor: Consolidate JSON parser tests into Configuration/Json directory
Move JSON parser test files to match source code organization:
- Move JsonConfigParserTests.cs from Tests/Json to Tests/Configuration/Json
- Move JsonResponseParserTests.cs from Tests/Configuration to Tests/Configuration/Json
- Update namespaces to Onyx.Tanuki.Tests.Configuration.Json

This aligns test organization with the JSON parser source files
located at Tanuki.Runtime/Configuration/Json/.
