+++ src/Tanuki.Tests/Configuration/Json/JsonConfigParserTests.cs
@@ -5,7 +5,7 @@ using Onyx.Tanuki.Configuration.Exceptions;
 using Onyx.Tanuki.Configuration.Json;
 using Xunit;
 
-namespace Onyx.Tanuki.Tests.Json;
+namespace Onyx.Tanuki.Tests.Configuration.Json;
 
 public class JsonConfigParserTests
 {
